Reusable RequestGroup.add_{resource|trait}
This simply refactors the (minimal) logic to add a resource or
(required/forbidden) trait to a RequestGroup into RequestGroup itself.
It didn't make sense for it to be in ResourceRequest.

This will make the logic reusable for e.g. assembling
RequestSpec.requested_resources for accelerator requests.

This is a straight refactor, thoroughly covered by existing tests.

Change-Id: Ied2047a6bd22d194b3c33dca44bf81e7eebc6c86