Move _get_request_group_mapping() to RequestSpec
The internals of the static ComputeManager._get_request_group_mapping
call was moved to RequestSpec as a member function. This will help in
subsequent patches where such call needs to be used from outside of the
ComputeManager.

This patch also adds unit test coverage for the moved internals of that
function.

Change-Id: I41868f8dc0aa5bff35d0e9e9982791f99df6fe0d
blueprint: support-move-ops-with-qos-ports-ussuri