Tie requester_id to RequestGroup suffix
Previously, request groups built from ports would have a unique integer
suffix generated for them as the suffix used for the GET
/allocation_candidates query to placement. Since [1] we've been using a
microversion permitting string suffixes (added in 1.33). This allows us
to use the port UUID (pulled from the RequestGroup.requester_id) as the
suffix instead.

This lays the groundwork for future patches that will:
- Take advantage of the suffix-to-provider mappings returned by GET
  /allocation_candidates as of placement microversion 1.34 (also
  included in [1]) to get rid of the
  map_requested_resources_to_providers hack.
- Seamlessly support request groups from other sources, such as cyborg
  device profiles.

[1] I52499ff6639c1a5815a8557b22dd33106dcc386b

Change-Id: I55e4d61ba1c5920c44fd71eb36a9ed2cc64c4117