Add QoS tempest config so bw tests run
In [1] the tempest-slow-py3 job was dropped and non-redundant bits
folded into the nova-next job.

Except we forgot to move over some of the config necessary to make this
QoS bandwidth test [2] work, so it gets skipped:

setUpClass (tempest.scenario.test_minbw_allocation_placement.MinBwAllocationPlacementTest)
... SKIPPED: Skipped as no physnet is available in config for placement based QoS allocation.

This syncs the nova-next job with the config like what was done for
tempest-slow here [3]. Some of that was already in place (to make
QoS heal allocation testing work).

[1] https://review.opendev.org/#/c/683988/
[2] https://github.com/openstack/tempest/blob/3eb3c29e979fd3f13c205d62119748952d63054a/tempest/scenario/test_minbw_allocation_placement.py#L142
[3] https://github.com/openstack/tempest/commit/c87a06b3c29427dc8f2513047c804e0410b4b99c

Closes-Bug: #1854993

Change-Id: Ib9332b8677c4ccb9e38b6fcdf39e989a891a56fe