Skip test_minbw_allocation_placement in nova-next job
We just recently enable this scenario test [1] but it's
causing the nova-next job to intermittently fail because
the nova-next job runs the tempest compute API and scenario
tests concurrently (unlike the tempest-full job which runs
the API tests concurrently and then the scenario tests serially)
and the scenario test is creating a public shared network which
causes the compute API tests, when creating a server without
specifying a network, to fail since nova doesn't know which
network to use. So for now just blacklist the scenario test
until the related tempest bug is fixed.

[1] https://review.opendev.org/#/c/697180/

Change-Id: Ie0e6136e88bedb2c4c6d0b742319627d58da147e
Related-Bug: #1844568