Remove dict compat from populate_filter_properties
The only callers of populate_filter_properties are conductor
methods and they all pass a Selection object so we can fulfill
the TODO and remove the dict compat. Note that the callers have
to pass a Selection object because they ask select_destinations
to return_objects=True and if there was a pinned RPC version the
select_destinations call would fail with
SelectionObjectsWithOldRPCVersionNotSupported. Point being, we
don't need to worry about old RPC compat code passing a dict rather
than a Selection object.

Change-Id: I076b84d46e5b2ff3bb6ed4d0c361683097c4c014