FUP for in-place numa rebuild
This patch addresses a number of typos and minor
issues raised during review of [1][2]. A summary
of the changes are corrections to typos in comments,
a correction to the exception message, an update to
the release note and the addition of debug logging.

[1] I0322d872bdff68936033a6f5a54e8296a6fb3434
[2] I48bccc4b9adcac3c7a3e42769c11fdeb8f6fd132

Related-Bug: #1804502
Related-Bug: #1763766

Conflicts:
    nova/tests/functional/libvirt/test_numa_servers.py
NOTE(sean-k-mooney): conflict was due to the use of
NUMAHostInfo instead of HostInfo.

Change-Id: I8975e524cd5a9c7dfb065bb2dc8ceb03f1b89e7b
(cherry picked from commit f6060ab6b54261ff50b8068732f6e509619d713e)
(cherry picked from commit 48bb9a9663374936221144bb6a24688128a51146)