[scheduler] Add a filter for vcenter shards
We need those shards to scale further in an availability-zone. We shard
on project level by adding the appropriate tags to the project in
keystone. If the project tags from keystone contain the tag
"sharding_enabled" then the hosts in _all_ shards will pass the shard
filter for this project.

This filter should only allow a host to pass if it's in an aggregate
having the requesting project's shard assigned.

Instead of using an attribute on any aggregate for a vcenter shard, we
use the name of the aggregate instead. It doesn't really make sense to
use the attribute, because we can't use an aggregate for multiple AZs
shards, because otherwise a project's ('vc-a-1', 'vc-b-0') would match
any host in a ('vc-a-0', 'vc-b-0')-aggregate for a deployment to the 'a'
availability zone.

`sharding_enabled` was introduced to facilitate both enabling sharding
(only one simple tag to set), and mainly for frontend code to detect
sharding status (mostly) without parsing tag strings. (If sharding is
not enabled, then vc-* tags will have to be parsed to find out which
shard(s) the project is on.)

We don't expect ironic nodes to be in a vc-* host aggregate, therefore
we log it to debug instead of error for such nodes.

Co-Authored-By: Jakob Karge <jakob.karge@sap.com>
Co-Authored-By: Marius Leustean <marius.leustean@sap.com>
Co-Authored-By: Fabian Wiesel <fabian.wiesel@sap.com>
(cherry picked from commit 1c19720b29a8e805a1f6ea51ea45e2638586ca8f)
(cherry picked from commit 56b0aabaca2097b13ebb9b4bcd3d94f222b1f369)
Incorporated-Change-Id: Iae29c9dd9728528fedc57e70094721cb918c36a6
Incorporated-Change-Id: I60f8a170cfb76235d6a9f77b53f9573331972c5a
Incorporated-Change-Id: I9f9b2c4f3eae642d78fb349f5c752711bcd94af2
Partly-Incorporated-Change-Id: I9d6f715faecc6782f93a3cd7f83f85f5ece02e60

Change-Id: I3d3b0601dc939b55ce0e956342f51ec3ea987888