Func test for qos live migration reschedule
This patch adds extra functional test coverage for live migration with
qos ports. It covers live migration with target host specified and
re-schedule success and failure cases.

There are two non test code changes.

1)
In the check_can_live_migrate_destination compute manager method CONF.host
was used to determine the destination host. This was replaced with
self.host. This change result in identical data in a production
environment. But during the functional test the CONF object is global for
every compute servers running in the environment. This causes that when
the above call runs on host2 the CONF.host is already set to host3 as
that host was started later in the test. Fortunately self.host is
set on the compute manage during the compute startup so it is correct in
the functional test env too. It is unfortunate that we have to change
non test code to make the test code work but the global CONF variable is
a hard problem to resolve.

2)
The InstancePCIRequest is saved in the LiveMigrationTask as the task
updates the request for each host it tries to move the instance. If non
of the destination supports the migration then the rollback() of the
task will restore the InstancePCIRequest to its original value as the
instance will remain on the source host.

Change-Id: I351399f02d4d3b3e958a62fe37380577f3056d0d
blueprint: support-move-ops-with-qos-ports-ussuri