Add recreate test for bug 1855927
When _poll_unconfirmed_resizes runs, if the actual resize
confirm fails on the source host the migration status may
be stuck in "confirming" status if it never reached the
source compute. Subsequent runs of _poll_unconfirmed_resizes
will not be able to auto-confirm the resize nor will the user
be able to manually confirm the resize. An admin could reset
the status on the server to ACTIVE or ERROR but that means
the source compute never gets cleaned up since you can only
confirm or revert a resize on a server with VERIFY_RESIZE status.

This adds a functional recreate test for the bug.

Note that if the request does get to the source compute and fails,
the _error_out_instance_on_exception context manager should set
the instance to ERROR status and the errors_out_migration decorator
should set the migration status to "error". The problem is if we
never get that far we can't recover automatically or via the API
without operator intervention.

Change-Id: I93e27d907a237ae16b6b4c7202ad157b067aa30e
Related-Bug: #1855927