Move common test method up to base class
To be able to test cancelling a live migration we need a special virt
driver in the functional test which requires a separate test class to
use. Before this separate test class is added this patch pulls up some
of the common test methods to an already existing base class so those
can be used from the new test class in a subsequent patch.

Change-Id: I83937bc8a4d7665c7f435020e284e269b98985c8
blueprint: support-move-ops-with-qos-ports-ussuri