pike-only: remove broken non-voting ceph jobs
The ceph job is broken on pike and the dependent
fix in devstack [1] has not gotten attention after
several months.

Since pike is in extended maintenance mode, the job
is non-voting even if it were working, and there does
not seem to be much desire in making this work again
(it would still be non-voting and could easily break
again), we might as well just remove it from pike
testing.

Related mailing list thread: [2]

[1] https://review.opendev.org/684756/
[2] http://lists.openstack.org/pipermail/openstack-discuss/2019-December/011632.html

Change-Id: I9e153f86c81ed6d9f8d9682b66d6d5c7f7b25296
Closes-Bug: #1835627