ocata-only: drop non-voting ceph job
Nova on stable/ocata is the last remaining user of
legacy-tempest-dsvm-full-devstack-plugin-ceph and
runs it non-voting. stable/ocata is in extended
maintenance mode and doesn't see much action anymore.
Furthermore, the ceph job is being removed from nova
in stable/pike testing [1] so it'd be a bit weird to
run a ceph job on ocata but not pike so let's just
remove it. Since it's non-voting if it breaks no one
will likely notice anyway.

[1] I9e153f86c81ed6d9f8d9682b66d6d5c7f7b25296

Change-Id: I3b088925dd46898bda6f74af4b933cfa732c31b7