Fix excessive runtime of test test_migrate_within_cell
The test test_migrate_within_cell functional test takes 160 seconds to
run.

Seems like the cleanup after the test function takes the extra time.
By looking at the test case it starts a migration then if that call
returns 202 the test finishes. So the cleanup and the migration happen
in parallel. By adding a wait for the server to finish migration
removes the excessive run time.

Change-Id: Ib4e1712d9724147ec8bdf90a1294041e62dfb136
Closes-Bug: #1858639