libvirt: Support boot from volume stable device instance rescue
This change introduces support for boot from volume stable device
instance rescue to the Libvirt driver by reporting the
supports_bfv_rescue capability.

The only change to the rescue method within the Libvirt driver is the
introduction of a call to utils.get_bdm_image_metadata when we fail to
lookup the image metadata of an instance. This function allowing us to
extract image metadata if any is present from the provided
block_device_mapping.

Implements: blueprint virt-bfv-instance-rescue
Change-Id: I99194539bd77790896a79fee47978905cba99bee