Add NovaEphemeralObject class for non-persistent objects
This adds a NovaEphemeralObject class, which inherits from NovaObject
and also from EphemeralObject. The latter is proposed against o.vo,
and is copied here for expediency.

In the past, all of our objects were representations of rows of columns
in the database and thus it was very important to not default fields
in objects when working on a record remotely that may have values in
the database for those fields which we did not wish to overwrite. Lately,
we have a lot of objects that are used purely over RPC, or which are stored
in the database as a blob and for which the explicit defaulting nature
is annoying. This is an attempt to classify such objects in a single
hierarchy and allow them to behave differently.

We already have some objects in the tree which do their own defaulting
in __init__, against what should be our policy. Some of those can be
made to inherit from the new class and continue, while others need to
be fixed. This patch adds a test to ensure that things comply with the
policy henceforth. Objects that had to be changed in the same patch as
the test are fixed here.

Specific object changes are:

- Diagnostics: Not persisted, can be converted to ephemeral
- InstanceNUMACell: Stored as a blob, can be converted to ephemeral.
                    Removes a test that asserted no dirty fields left
                    after init, but for no reason and not compatible
                    with the base class.
- PCIDevice: Persited row-based, so cannot be ephemeral. Converts it
             to set default for the one field on lazy-load
- Quotas, QuotasNoOp: Stored row-based, so cannot be ephemeral.
                      Converts it to set default on lazy-load,
                      and persists the resetting of dirty fields for
                      compatibility with what is there (even though it
                      is probably not needed).
- RequestGroup: Not persisted, can be converted to ephemeral.
- Service: Does defaulting in init, but specifically because we *want*
           to always overwrite what is in the database on save. This is
           part of how service version works to always know the version of
           compute (et al) services once they start up.

For the objects that were converted to Ephemeral, some hashes changed but
only because defaults were added, so no need to bump the version, just update
the hash. The hash checker is trying to make sure we bump the version when
necessary, but in this case, a default does not affect RPC or object behavior
in an incompatible way, so no bump is needed.

Change-Id: Ica9f217d0318fc7c2db4bcdea12d00aad749c30c