Clarify fitting hugepages log message
Currently, when fitting an instance with hugepages to a host, if the
host NUMA cell does not have enough free pages of the requested size,
the message that is logged is about not supporting the page size at
all. This is not correct. This patch changes the error message to
mention the possibility that there are not enough free pages of the
requested size on the host.

Change-Id: Ifcba63b654fcdce279e93f3b726aff09a48807ed