functional: Stop setting Flavor.id
In change I475ea0fa5f2d5b197118f0ced5a0ff6907411972, we changed how we
generated flavor names and IDs to stop basing them on existing flavors.
However, the call to 'randint(10000)' that we used instead has proven
problematic since there is a high chance of collisions that will only
increase as the number of tests increase. We could switch back to the
previous scheme but that's unnecessary since there's actually no reason
we need to set 'Flavor.id' in the first place...so don't.

A now-unused remnant of the "old way" is also removed, since it was
spotted while fixing this.

Change-Id: Iab6245bc5ed8f95dae9c384b96e6bef0add7dca6
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>
Closes-bug: #1860417